oauth_example/authorization_code_flow/pyproject.toml
[tool.poetry]
name = "authorization_code_flow"
version = "0.1.0"
description = "Authorization code flow example"
authors = ["Mohamed Nesredin <mohamednesredin19@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "3.8.5"
fastapi = "^0.60.1"
pendulum = "^2.1.2"
pyjwt = "^1.7.1"
python-dotenv = "^0.14.0"
tortoise-orm = "^0.16.14"
httpx = "^0.13.3"
uvicorn = "^0.11.6"
itsdangerous = "^1.1.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
